+++ src/Exception/RuntimeException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Pagerfanta\Exception;
+
+class RuntimeException extends \RuntimeException implements Exception
+{
+}