+++ src/View/Template/TwitterBootstrapTemplate.php
@@ -16,6 +16,9 @@ namespace Pagerfanta\View\Template;
  */
 class TwitterBootstrapTemplate extends Template
 {
+    /**
+     * @var string[]
+     */
     protected static $defaultOptions = [
         'prev_message' => '&larr; Previous',
         'next_message' => 'Next &rarr;',
@@ -31,105 +34,87 @@ class TwitterBootstrapTemplate extends Template
         'rel_next' => 'next',
     ];
 
-    public function container()
+    public function container(): string
     {
         return sprintf('<div class="%s"><ul>%%pages%%</ul></div>',
             $this->option('css_container_class')
         );
     }
 
-    public function page($page)
+    public function page($page): string
     {
-        $text = $page;
-
-        return $this->pageWithText($page, $text);
+        return $this->pageWithText($page, $page);
     }
 
-    public function pageWithText($page, $text)
+    public function pageWithText($page, $text, ?string $rel = null): string
     {
-        $class = null;
-
-        return $this->pageWithTextAndClass($page, $text, $class);
+        return $this->pageWithTextAndClass($page, $text, '', $rel);
     }
 
-    private function pageWithTextAndClass($page, $text, $class, $rel = null)
+    private function pageWithTextAndClass($page, $text, $class, ?string $rel = null): string
     {
-        $href = $this->generateRoute($page);
-
-        return $this->linkLi($class, $href, $text, $rel);
+        return $this->linkLi($class, $this->generateRoute($page), $text, $rel);
     }
 
-    public function previousDisabled()
+    public function previousDisabled(): string
     {
-        $class = $this->previousDisabledClass();
-        $text = $this->option('prev_message');
-
-        return $this->spanLi($class, $text);
+        return $this->spanLi($this->previousDisabledClass(), $this->option('prev_message'));
     }
 
-    private function previousDisabledClass()
+    private function previousDisabledClass(): string
     {
         return $this->option('css_prev_class').' '.$this->option('css_disabled_class');
     }
 
-    public function previousEnabled($page)
+    public function previousEnabled($page): string
     {
-        $text = $this->option('prev_message');
-        $class = $this->option('css_prev_class');
-        $rel = $this->option('rel_previous');
-
-        return $this->pageWithTextAndClass($page, $text, $class, $rel);
+        return $this->pageWithTextAndClass($page, $this->option('prev_message'), $this->option('css_prev_class'), $this->option('rel_previous'));
     }
 
     public function nextDisabled()
     {
-        $class = $this->nextDisabledClass();
-        $text = $this->option('next_message');
-
-        return $this->spanLi($class, $text);
+        return $this->spanLi($this->nextDisabledClass(), $this->option('next_message'));
     }
 
-    private function nextDisabledClass()
+    private function nextDisabledClass(): string
     {
         return $this->option('css_next_class').' '.$this->option('css_disabled_class');
     }
 
-    public function nextEnabled($page)
+    public function nextEnabled($page): string
     {
-        $text = $this->option('next_message');
-        $class = $this->option('css_next_class');
-        $rel = $this->option('rel_next');
-
-        return $this->pageWithTextAndClass($page, $text, $class, $rel);
+        return $this->pageWithTextAndClass($page, $this->option('next_message'), $this->option('css_next_class'), $this->option('rel_next'));
     }
 
-    public function first()
+    public function first(): string
     {
         return $this->page(1);
     }
 
-    public function last($page)
+    public function last($page): string
     {
         return $this->page($page);
     }
 
-    public function current($page)
+    public function current($page): string
     {
         $text = trim($page.' '.$this->option('active_suffix'));
-        $class = $this->option('css_active_class');
 
-        return $this->spanLi($class, $text);
+        return $this->spanLi($this->option('css_active_class'), $text);
     }
 
-    public function separator()
+    public function separator(): string
     {
-        $class = $this->option('css_dots_class');
-        $text = $this->option('dots_message');
-
-        return $this->spanLi($class, $text);
+        return $this->spanLi($this->option('css_dots_class'), $this->option('dots_message'));
     }
 
-    protected function linkLi($class, $href, $text, $rel = null)
+    /**
+     * @param string $class
+     * @param string $href
+     * @param int|string $text
+     * @param string|null $rel
+     */
+    protected function linkLi($class, $href, $text, $rel = null): string
     {
         $liClass = $class ? sprintf(' class="%s"', $class) : '';
         $rel = $rel ? sprintf(' rel="%s"', $rel) : '';
@@ -137,7 +122,11 @@ class TwitterBootstrapTemplate extends Template
         return sprintf('<li%s><a href="%s"%s>%s</a></li>', $liClass, $href, $rel, $text);
     }
 
-    protected function spanLi($class, $text)
+    /**
+     * @param string $class
+     * @param int|string $text
+     */
+    protected function spanLi($class, $text): string
     {
         $liClass = $class ? sprintf(' class="%s"', $class) : '';